Introduce error batch plugin
