+++ src/ServiceControl/Monitoring/HeartbeatMonitoringComponent.cs
@@ -49,6 +49,7 @@
                 collection.AddIntegrationEventPublisher<HeartbeatStoppedPublisher>();
 
                 collection.AddErrorMessageEnricher<DetectNewEndpointsFromErrorImportsEnricher>();
+                collection.AddErrorMessageBatchPlugin<SaveNewEndpointsErrorMessageBatchPlugin>();
 
                 collection.AddPlatformConnectionProvider<HeartbeatsPlatformConnectionDetailsProvider>();
             });
