+++ src/ServiceControl/Operations/ErrorProcessor.cs
@@ -10,7 +10,6 @@
     using Infrastructure.DomainEvents;
     using Infrastructure.Metrics;
     using MessageFailures;
-    using Monitoring;
     using NServiceBus;
     using NServiceBus.Logging;
     using NServiceBus.Transport;
@@ -34,22 +33,16 @@
                                         ""Raven-Entity-Name"": ""{FailedMessage.CollectionName}"",
                                         ""Raven-Clr-Type"": ""{typeof(FailedMessage).AssemblyQualifiedName}""
                                     }}");
-
-            KnownEndpointMetadata = RavenJObject.Parse($@"
-                                    {{
-                                        ""Raven-Entity-Name"": ""{KnownEndpoint.CollectionName}"",
-                                        ""Raven-Clr-Type"": ""{typeof(KnownEndpoint).AssemblyQualifiedName}""
-                                    }}");
         }
 
         public ErrorProcessor(BodyStorageEnricher bodyStorageEnricher, IEnrichImportedErrorMessages[] enrichers, IFailedMessageEnricher[] failedMessageEnrichers, IDomainEvents domainEvents,
-            Counter ingestedCounter, EndpointInstanceMonitoring endpointInstanceMonitoring)
+            Counter ingestedCounter, IErrorMessageBatchPlugin[] batchPlugins)
         {
             this.bodyStorageEnricher = bodyStorageEnricher;
             this.enrichers = enrichers;
             this.domainEvents = domainEvents;
             this.ingestedCounter = ingestedCounter;
-            this.endpointInstanceMonitoring = endpointInstanceMonitoring;
+            this.batchPlugins = batchPlugins;
             failedMessageFactory = new FailedMessageFactory(failedMessageEnrichers);
         }
 
@@ -77,7 +70,10 @@
                 ingestedCounter.Mark();
             }
 
-            AddKnownEndpoints(storedContexts, commands);
+            foreach (var batchPlugin in batchPlugins)
+            {
+                batchPlugin.AfterProcessing(contexts, commands);
+            }
 
             return (storedContexts, commands);
         }
@@ -107,15 +103,6 @@
                 FailedMessageId = headers.UniqueId()
             });
         }
-
-        static PutCommandData CreateKnownEndpointsPutCommand(KnownEndpoint endpoint) =>
-            new PutCommandData
-            {
-                Document = RavenJObject.FromObject(endpoint),
-                Etag = null,
-                Key = endpoint.Id.ToString(),
-                Metadata = KnownEndpointMetadata
-            };
 
         async Task ProcessMessage(MessageContext context)
         {
@@ -223,64 +210,13 @@
             };
         }
 
-        void AddKnownEndpoints(List<MessageContext> batch, ICollection<ICommandData> commands)
-        {
-            var knownEndpoints = new Dictionary<string, KnownEndpoint>();
-            foreach (var context in batch)
-            {
-                var errorEnricherContext = context.Extensions.Get<ErrorEnricherContext>();
-
-                void RecordKnownEndpoint(string key)
-                {
-                    if (errorEnricherContext.Metadata.TryGetValue(key, out var endpointObject)
-                        && endpointObject is EndpointDetails endpointDetails
-                        && endpointDetails.HostId != Guid.Empty) // for backwards compat with version before 4_5 we might not have a hostid
-                    {
-                        if (endpointInstanceMonitoring.IsNewInstance(endpointDetails))
-                        {
-                            RecordKnownEndpoints(endpointDetails, knownEndpoints);
-                        }
-                    }
-                }
-
-                RecordKnownEndpoint("SendingEndpoint");
-                RecordKnownEndpoint("ReceivingEndpoint");
-            }
-
-            foreach (var endpoint in knownEndpoints.Values)
-            {
-                if (Logger.IsDebugEnabled)
-                {
-                    Logger.Debug($"Adding known endpoint '{endpoint.EndpointDetails.Name}' for bulk storage");
-                }
-
-                commands.Add(CreateKnownEndpointsPutCommand(endpoint));
-            }
-        }
-
-        static void RecordKnownEndpoints(EndpointDetails observedEndpoint, Dictionary<string, KnownEndpoint> observedEndpoints)
-        {
-            var uniqueEndpointId = $"{observedEndpoint.Name}{observedEndpoint.HostId}";
-            if (!observedEndpoints.TryGetValue(uniqueEndpointId, out KnownEndpoint _))
-            {
-                observedEndpoints.Add(uniqueEndpointId, new KnownEndpoint
-                {
-                    Id = DeterministicGuid.MakeId(observedEndpoint.Name, observedEndpoint.HostId.ToString()),
-                    EndpointDetails = observedEndpoint,
-                    HostDisplayName = observedEndpoint.Host,
-                    Monitored = false
-                });
-            }
-        }
-
         readonly IEnrichImportedErrorMessages[] enrichers;
         readonly IDomainEvents domainEvents;
         readonly Counter ingestedCounter;
+        readonly IErrorMessageBatchPlugin[] batchPlugins;
         BodyStorageEnricher bodyStorageEnricher;
         FailedMessageFactory failedMessageFactory;
-        readonly EndpointInstanceMonitoring endpointInstanceMonitoring;
         static readonly RavenJObject FailedMessageMetadata;
-        static readonly RavenJObject KnownEndpointMetadata;
         static readonly JsonSerializer Serializer;
         static readonly ILog Logger = LogManager.GetLogger<ErrorProcessor>();
     }
